backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
alembic
psycopg2-binary
python-dotenv
langchain
langchain-google-genai
google-generativeai
python-multipart
pypdfium2
python-docx